rustdoc: clean up search query parser
This commit simplifies the parser by:

* Separating lexing from parsing
* Using a more principled pratt parser for precedence,
  based on the approach described in [matklad's blog]

The search.js file actually gets smaller, since parsing is driven by an
operator table and lexing is driven by a regex.

Most of the test changes are inconsequential changes in error messages
for some of the more contrived test cases. They're different, but not
really worse, and error handling code is easier to write because it can
look at lexical tokens instead of raw text.

The one major change in behavior is for this query:

    a:: b

The old parser rejected it, but this one allows it. This change brings
search.js closer to rustc, and makes more sense: the old behavior made
more sense back when space-separated paths were treated as function
parameters, but since b99bf4677bb157ad46758099189389ac8cc60779 changed
it so that space separated idents form paths, treating it as equivalent
to `a::b` is at least equally justifiable.

[matklad's blog]: https://matklad.github.io/2020/04/13/simple-but-powerful-pratt-parsing.html